Create new Express and Angular project
A 'backend' and 'frontend' directory was created containing an express
and angular project respectively. The decision was made based on best
practices and for future scalability.

Additionally, a README.md file was added for describing key features
about our project, its implementation, and how you can run it locally on
you machine.